+++ reader.h
@@ -147,7 +147,7 @@ void reader_set_complete_function( complete_function_t );
  The type of a highlight function.
  */
 class env_vars_snapshot_t;
-typedef void (*highlight_function_t)( const wcstring &, std::vector<int> &, int, wcstring_list_t *, const env_vars_snapshot_t &vars );
+typedef void (*highlight_function_t)( const wcstring &, std::vector<int> &, size_t, wcstring_list_t *, const env_vars_snapshot_t &vars );
 
 /**
  Specify function for syntax highlighting. The function must take these arguments:
